Moving DocComment/DocNode into separate files in an "api/nodes" subfolder
