+++ tsdoc/scripts/deploy.js
@@ -1,0 +1,4 @@
+console.log('Executing common/scripts/deploy.js');
+console.log('ARGV: ' + JSON.stringify(process.argv));
+console.log('CWD: ' + process.cwd());
+console.log('NPM_INITCWD: ' + process.env['INIT_CWD']);
