+++ tsdoc/src/api/nodes/DocComment.ts
@@ -1,28 +1,6 @@
-import { TextRange } from './TextRange';
-import { ParseError } from './ParseError';
-
-/**
- * Indicates the type of {@link DocNode}.
- */
-export enum DocNodeKind {
-  DocComment
-}
-
-/**
- * Represents a parsed documentation item.
- */
-export abstract class DocNode {
-  /**
-   * Indicates the kind of DocNode.
-   */
-  public abstract readonly kind: DocNodeKind;
-
-  public readonly range: TextRange;
-
-  public constructor(range: TextRange) {
-    this.range = range;
-  }
-}
+import { TextRange } from '../TextRange';
+import { ParseError } from '../ParseError';
+import { DocNode, DocNodeKind } from './DocNode';
 
 /**
  * Constructor parameters for {@link DocComment}.
