+++ tsdoc/src/index.ts
@@ -1,12 +1,6 @@
-export {
-  DocNodeKind,
-  DocComment,
-  DocNode
-} from './api/DocComment';
+
+export { DocNodeKind, DocNode } from './api/nodes/DocNode';
+export { DocComment } from './api/nodes/DocComment';
 
 export { TSDocParser } from './api/TSDocParser';
-
-export {
-  TextRange,
-  ITextLocation
-} from './api/TextRange';
+export { TextRange, ITextLocation } from './api/TextRange';
